Moved VerificationData to public package
Needed for #719
