+++ src/main/java/org/mockito/internal/verification/MockAwareVerificationMode.java
@@ -4,7 +4,7 @@
  */
 package org.mockito.internal.verification;
 
-import org.mockito.internal.verification.api.VerificationData;
+import org.mockito.verification.VerificationData;
 import org.mockito.verification.VerificationMode;
 
 public class MockAwareVerificationMode implements VerificationMode {
