+++ src/main/java/org/mockito/internal/verification/Only.java
@@ -13,7 +13,7 @@ import static org.mockito.internal.invocation.InvocationsFinder.findInvocations;
 import java.util.List;
 
 import org.mockito.internal.invocation.InvocationMatcher;
-import org.mockito.internal.verification.api.VerificationData;
+import org.mockito.verification.VerificationData;
 import org.mockito.invocation.Invocation;
 import org.mockito.verification.VerificationMode;
