+++ src/test/java/org/mockito/internal/verification/OnlyTest.java
@@ -8,7 +8,7 @@ import org.junit.Test;
 import org.mockito.exceptions.base.MockitoAssertionError;
 import org.mockito.internal.invocation.InvocationBuilder;
 import org.mockito.internal.invocation.InvocationMatcher;
-import org.mockito.internal.verification.api.VerificationData;
+import org.mockito.verification.VerificationData;
 import org.mockito.invocation.Invocation;
 
 import java.util.Arrays;
